Nmap/requirements.txt
python-libnmap
networkx